+++ lib/minitest/event_prof_formatter.rb
@@ -11,35 +11,37 @@ module Minitest
       using ::TestProf::StringTruncate
       using ::TestProf::StringStripHeredoc
 
-      def initialize(profiler)
-        @profiler = profiler
+      def initialize(profilers)
+        @profilers = profilers
         @results = []
       end
 
       def prepare_results
-        total_results
-        by_groups
-        by_examples
+        @profilers.each do |profiler|
+          total_results(profiler)
+          by_groups(profiler)
+          by_examples(profiler)
+        end
         @results.join
       end
 
       private
 
-      def total_results
+      def total_results(profiler)
         @results <<
           <<-MSG.strip_heredoc
-            EventProf results for #{@profiler.event}
+            EventProf results for #{profiler.event}
 
-            Total time: #{@profiler.total_time.duration}
-            Total events: #{@profiler.total_count}
+            Total time: #{profiler.total_time.duration}
+            Total events: #{profiler.total_count}
 
-            Top #{@profiler.top_count} slowest suites (by #{@profiler.rank_by}):
+            Top #{profiler.top_count} slowest suites (by #{profiler.rank_by}):
 
           MSG
       end
 
-      def by_groups
-        @profiler.results[:groups].each do |group|
+      def by_groups(profiler)
+        profiler.results[:groups].each do |group|
           description = group[:id][:name]
           location = group[:id][:location]
 
@@ -50,11 +52,11 @@ module Minitest
         end
       end
 
-      def by_examples
-        return unless @profiler.results[:examples]
-        @results << "\nTop #{@profiler.top_count} slowest tests (by #{@profiler.rank_by}):\n\n"
+      def by_examples(profiler)
+        return unless profiler.results[:examples]
+        @results << "\nTop #{profiler.top_count} slowest tests (by #{profiler.rank_by}):\n\n"
 
-        @profiler.results[:examples].each do |example|
+        profiler.results[:examples].each do |example|
           description = example[:id][:name]
           location = example[:id][:location]